Changed some ints to size_ts

+++ LibHR/Geometry/hr_sendrecv.c
@@ -42,15 +42,15 @@ void hr_sendrecv(void *sendbuffer, void *recvbuffer, geometry_descriptor *type, 
     _BUFFER_FOR(i, nbuffers) {
         char *recv_buffer = (_GET_RECV_BUFFER((char *)recvbuffer, i, field_dim, type, chars_per_site));
         int recv_proc = type->rbuf_from_proc[i];
-        int number_of_sites = roundUp(type->rbuf_len[i], THREADSIZE) / 2;
-        int recv_size_in_dbl = field_dim * number_of_sites * mpi_chunks_per_site;
+        size_t number_of_sites = roundUp(type->rbuf_len[i], THREADSIZE) / 2;
+        size_t recv_size_in_dbl = field_dim * number_of_sites * mpi_chunks_per_site;
 
-        const int n_messages = (recv_size_in_dbl - 1) / MAX_MSG_SIZE + 1;
-        for (int msg_id = 0; msg_id < n_messages; msg_id++) {
+        const size_t n_messages = (recv_size_in_dbl - 1) / MAX_MSG_SIZE + 1;
+        for (size_t msg_id = 0; msg_id < n_messages; msg_id++) {
             char *recv_msg = recv_buffer + msg_id * MAX_MSG_SIZE * size_of_real;
-            int start_of_message = MAX_MSG_SIZE * msg_id;
-            int end_of_message = min(MAX_MSG_SIZE * (msg_id + 1), recv_size_in_dbl);
-            const int msg_size = end_of_message - start_of_message;
+            size_t start_of_message = MAX_MSG_SIZE * msg_id;
+            size_t end_of_message = min(MAX_MSG_SIZE * (msg_id + 1), recv_size_in_dbl);
+            const int msg_size = (int)(end_of_message - start_of_message);
             CHECK_MPI(
                 MPI_Irecv(recv_msg, msg_size, mpi_real_type, recv_proc, tag_counter, cart_comm, &(field_reqs[tag_counter])));
             tag_counter++;
@@ -63,15 +63,15 @@ void hr_sendrecv(void *sendbuffer, void *recvbuffer, geometry_descriptor *type, 
     _BUFFER_FOR(i, nbuffers) {
         char *send_buffer = (_GET_SEND_BUFFER((char *)sendbuffer, i, field_dim, type, chars_per_site));
         int send_proc = type->sbuf_to_proc[i];
-        int number_of_sites = roundUp(type->sbuf_len[i], THREADSIZE) / 2;
-        int send_size_in_dbl = field_dim * number_of_sites * mpi_chunks_per_site;
+        size_t number_of_sites = roundUp(type->sbuf_len[i], THREADSIZE) / 2;
+        size_t send_size_in_dbl = field_dim * number_of_sites * mpi_chunks_per_site;
 
-        const int n_messages = (send_size_in_dbl - 1) / MAX_MSG_SIZE + 1;
-        for (int msg_id = 0; msg_id < n_messages; msg_id++) {
+        const size_t n_messages = (send_size_in_dbl - 1) / MAX_MSG_SIZE + 1;
+        for (size_t msg_id = 0; msg_id < n_messages; msg_id++) {
             char *send_msg = send_buffer + msg_id * MAX_MSG_SIZE * size_of_real;
-            int start_of_message = MAX_MSG_SIZE * msg_id;
-            int end_of_message = min(MAX_MSG_SIZE * (msg_id + 1), send_size_in_dbl);
-            const int msg_size = end_of_message - start_of_message;
+            size_t start_of_message = MAX_MSG_SIZE * msg_id;
+            size_t end_of_message = min(MAX_MSG_SIZE * (msg_id + 1), send_size_in_dbl);
+            const int msg_size = (int)(end_of_message - start_of_message);
             CHECK_MPI(MPI_Isend(send_msg, msg_size, mpi_real_type, send_proc, tag_counter, cart_comm,
                                 &(field_reqs[2 * tag_counter])));
             tag_counter++;
